apps/api/requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic-settings==2.8.1
pydantic_core==2.41.5
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3
sympy==1.13.3
SQLAlchemy==2.0.43
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
